Better(?) Logging
A bunch of hacks to make the log output of Airtime somewhat manageable. Some of the are really needed some of them are more of a half baked attempt to get some serious info out of Airtime. Experimental and to be finalized after we have monitoring and the whole shebang set up.